feat: 네이버 환율 Open API 기능

+++ src/main/java/com/nayoung/exchangerateopenapitest/api/dto/account/AccountDto.java
@@ -3,10 +3,12 @@ package com.nayoung.exchangerateopenapitest.api.dto.account;
 import com.nayoung.exchangerateopenapitest.domain.account.Currency;
 import lombok.Builder;
 
+import java.math.BigDecimal;
+
 @Builder
 public record AccountDto (
 	Long id,
-	Long money,
+	BigDecimal money,
 	Currency currency
 ) {
 }
